Break main.rs into different files for readability
