+++ src/config.rs
@@ -1,0 +1,7 @@
+pub const DEFAULT_BASE16_SHELL_PATH: &str = "~/.config/tinted-theming/shell";
+pub const BASE16_SHELL_PATH_ENV_VAR_NAME: &str = "BASE16_SHELL_PATH";
+pub const BASE16_SHELL_HOOKS_PATH_ENV_VAR_NAME: &str = "BASE16_SHELL_HOOKS_PATH";
+pub const BASE16_SHELL_THEME_NAME_PATH_ENV_VAR_NAME: &str = "BASE16_SHELL_THEME_NAME_PATH";
+pub const BASE16_SHELL_CONFIG_PATH_ENV_VAR_NAME: &str = "BASE16_CONFIG_PATH";
+pub const XDG_CONFIG_HOME_ENV_VAR: &str = "XDG_CONFIG_HOME";
+pub const HOME_ENV_VAR: &str = "HOME";
